// mkv-subtitle-converter/frontend/src/lib/types/ipc.ts
// This file was generated by [specta]. DO NOT EDIT.
// Run `pnpm run generate:types` to update it.

import { z } from 'zod';
// This file has been generated by Specta. Do not edit this file manually.
export const DirectoryStatsSchema = z.object({
  exists: z.boolean(),
  fileCount: z.number(),
  totalSizeBytes: z.number(),
  files: z.array(z.lazy(() => FileStatSchema))
});
export type DirectoryStats = z.infer<typeof DirectoryStatsSchema>;

export const FailedFileSchema = z.object({
  path: z.string(),
  reason: z.string()
});
export type FailedFile = z.infer<typeof FailedFileSchema>;

export const FileStatSchema = z.object({
  name: z.string(),
  sizeBytes: z.number()
});
export type FileStat = z.infer<typeof FileStatSchema>;

export const FolderReportStatusSchema = z.object({
  hasSuccess: z.boolean(),
  hasFailure: z.boolean()
});
export type FolderReportStatus = z.infer<typeof FolderReportStatusSchema>;

export const IpcPayloadDataSchema = z.union([
  z.object({
    type: z.literal('cancelled'),
    payload: z.string()
  }),
  z.object({
    type: z.literal('fileProcessed'),
    payload: z.object({
      processed: z.number(),
      converted: z.number(),
      fileCompleted: z.string().nullable(),
      rootDirectory: z.string().nullable()
    })
  }),
  z.object({
    type: z.literal('folderStatusUpdate'),
    payload: z.object({
      folder: z.string(),
      status: z.string()
    })
  }),
  z.object({
    type: z.literal('startedScanned'),
    payload: z.object({
      totalCount: z.number(),
      folderCounts: z.record(z.string(), z.number())
    })
  })
]);
export type IpcPayloadData = z.infer<typeof IpcPayloadDataSchema>;

export const SubtitleMetadataSchema = z.object({
  file: z.string(),
  language: z.string(),
  trackName: z.string(),
  codec: z.string(),
  sourceFile: z.string()
});
export type SubtitleMetadata = z.infer<typeof SubtitleMetadataSchema>;